requirements.txt
pytest==7.2.0
python-lambda-local== 0.1.13